+++ soc/riscv/CMakeLists.txt
@@ -1,5 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
+add_subdirectory(common)
+
 if(SOC_FAMILY)
   add_subdirectory(${SOC_FAMILY})
 else()
