+++ soc/riscv/common/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory_ifdef(CONFIG_SOC_FAMILY_RISCV_PRIVILEGED riscv-privileged)
