+++ soc/riscv/common/riscv-privileged/idle.c
