+++ soc/riscv/common/riscv-privileged/soc_common_irq.c
