+++ soc/riscv/litex-vexriscv/CMakeLists.txt
@@ -5,8 +5,8 @@
 #
 
 zephyr_sources(
-    ../riscv-privileged/common/soc_irq.S
-    ../riscv-privileged/common/vector.S
+    ../common/riscv-privileged/soc_irq.S
+    ../common/riscv-privileged/vector.S
 )
 
 set(SOC_LINKER_SCRIPT ${ZEPHYR_BASE}/include/zephyr/arch/riscv/common/linker.ld CACHE INTERNAL "")
