+++ soc/riscv/riscv-privileged/CMakeLists.txt
@@ -1,4 +1,3 @@
 # SPDX-License-Identifier: Apache-2.0
 
-add_subdirectory(common)
 add_subdirectory(${SOC_SERIES})
